refactor: 메소드 위치 변경
부모 아이디를 통해 조회하는 경우, 역할을 명확히 하기 위하여 PickUp이 아닌 ParentRepository에서 역할을 수행하도록 변경
