+++ BE/iDrop/src/main/java/ifive/idrop/repository/PickUpRepository.java
@@ -31,15 +31,4 @@ public class PickUpRepository {
     public void savePickUp(PickUp pick) {
         em.persist(pick);
     }
-
-    public List<PickUpInfo> findRunningPickInfo(Long parentId) {
-        String query = "SELECT pui\n" +
-                "FROM PickUpInfo pui\n" +
-                "JOIN Child c ON pui.id = c.id\n" +
-                "JOIN PickUp pu ON pui.id = pu.pickUpInfo.id\n" +
-                "WHERE c.parent.id =: parentId AND pu.startTime IS NOT NULL";
-        return em.createQuery(query, PickUpInfo.class)
-                .setParameter("parentId", parentId)
-                .getResultList();
-    }
 }
